Removed bugs identified by findBugs, including the Feature class which had implemented equals() method but not hashCode() method.
git-svn-id: http://svn.jcvi.org/ANNOTATION/RulesBasedAnnotation/trunk@32729 061605af-e407-0410-a945-9c8a22772c06

+++ src/org/jcvi/annotation/facts/HomologyHit.java
@@ -145,11 +145,11 @@ public class HomologyHit {
     }
 
     public double getHitPercentLength() {
-        return 100 * (hitEnd - hitStart) / hitLength;
+        return ((double) 100 * (hitEnd - hitStart))/(double) hitLength;
     }
 
     public double getQueryPercentLength() {
-        return 100 * (queryEnd - queryStart) / queryLength;
+        return ((double) 100 * (queryEnd - queryStart))/(double) queryLength;
     }
 
     public double getHitQueryLengthRatio() {
